+++ jib-cli/src/main/java/com/google/cloud/tools/jib/cli/jar/JarFiles.java
@@ -89,6 +89,9 @@ public class JarFiles {
     if (processor.getJavaVersion() <= 11) {
       return "eclipse-temurin:11-jre";
     }
-    return "eclipse-temurin:17-jre";
+    if (processor.getJavaVersion() <= 17) {
+      return "eclipse-temurin:17-jre";
+    }
+    return "eclipse-temurin:21-jre";
   }
 }